# MobyPark/OpenAPI/openapi.yaml
openapi: 3.0.4
info: {}
paths:
...
  /api/HotelPasses/parkinglot/{id}:
    get:
      tags:
      - HotelPasses
      summary: Retrieves all passes for a specific parking lot.
      parameters:
      - name: parkingLotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadHotelPassDto'
        '404':
          description: Parking lot has no passes
...
  /api/ParkingLots:
    post:
      tags:
      - ParkingLots
      summary: Creates a new parking lot.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParkingLotDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateParkingLotDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateParkingLotDto'
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Parking lot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadParkingLotDto'
        '409':
          description: Parking lot address already taken
  /api/ParkingLots/{lotId}:
    get:
      tags:
      - ParkingLots
      summary: Retrieves a parking lot by its ID.
      parameters:
      - name: lotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Parking lot found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadParkingLotDto'
        '404':
          description: Parking lot not found
  /api/ParkingLots/by-address:
    get:
      tags:
      - ParkingLots
      summary: Retrieves a parking lot by its address.
      parameters:
      - name: address
        in: query
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Parking lot found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadParkingLotDto'
        '404':
          description: Parking lot not found
    patch:
      tags:
      - ParkingLots
      summary: Updates a parking lot by address.
      parameters:
      - name: address
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
      responses:
        '400':
          description: Invalid update data
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadParkingLotDto'
        '404':
          description: Parking lot not found
        '409':
          description: New address conflicts with existing lot
  /api/ParkingLots/by-id/{lotId}:
    patch:
      tags:
      - ParkingLots
      summary: Updates a parking lot by ID.
      parameters:
      - name: lotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
          text/json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PatchParkingLotDto'
      responses:
        '400':
          description: Invalid update data
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadParkingLotDto'
        '404':
          description: Parking lot not found
        '409':
          description: New address conflicts with existing lot
    delete:
      tags:
      - ParkingLots
      summary: Deletes a parking lot by ID.
      parameters:
      - name: lotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Deletion successful
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Parking lot not found
  /api/ParkingLots/by-address/{address}:
    delete:
      tags:
      - ParkingLots
      summary: Deletes a parking lot by address.
      parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: Deletion successful
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Parking lot not found
  /api/ParkingLots/all:
    get:
      tags:
      - ParkingLots
      summary: Retrieves all registered parking lots.
      responses:
        '400':
          description: Invalid data supplied
        '401':
          description: Unauthorized
        '500':
          description: Unexpected internal server error
        '200':
          description: List retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadParkingLotDto'
...
components:
  schemas:
...
    CreateParkingLotDto:
      required:
      - address
      - capacity
      - location
      - name
      - tariff
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: The display name of the parking lot
        location:
          minLength: 1
          type: string
          description: The general location or city (e.g. 'Downtown')
        address:
          minLength: 1
          type: string
          description: The full street address
        capacity:
          maximum: 20000
          minimum: 1
          type: integer
          description: The total number of parking spots available
          format: int32
        tariff:
          maximum: 100
          minimum: 0
          type: number
          description: The standard hourly tariff
          format: double
        dayTariff:
          type: number
          description: Optional discounted tariff for full-day parking
          format: double
          nullable: true
      additionalProperties: false
      description: Data required to create a new parking lot.
...
    LicensePlateModel:
      required:
      - licensePlateNumber
      type: object
      properties:
        licensePlateNumber:
          minLength: 1
          type: string
      additionalProperties: false
...
    ParkingLotModel:
      required:
      - address
      - capacity
      - createdAt
      - location
      - name
      - reservations
      - reserved
      - status
      - tariff
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          minLength: 1
          type: string
        location:
          minLength: 1
          type: string
        address:
          minLength: 1
          type: string
        capacity:
          type: integer
          format: int32
        reserved:
          type: integer
          format: int32
        availableSpots:
          type: integer
          format: int32
          readOnly: true
        tariff:
          type: number
          format: double
        dayTariff:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ParkingLotStatus'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/ReservationModel'
      additionalProperties: false
    ParkingLotStatus:
      enum:
      - 0
      - 1
      - 2
      type: integer
      format: int32
...
    PatchParkingLotDto:
      type: object
      properties:
        name:
          type: string
          description: The new display name
          nullable: true
        location:
          type: string
          description: The new general location
          nullable: true
        address:
          type: string
          description: The new street address (must be unique)
          nullable: true
        capacity:
          maximum: 20000
          minimum: 1
          type: integer
          description: The new total capacity
          format: int32
          nullable: true
        tariff:
          maximum: 100
          minimum: 0
          type: number
          description: The new hourly tariff
          format: double
          nullable: true
        dayTariff:
          type: number
          description: The new day tariff
          format: double
          nullable: true
        reserved:
          maximum: 20000
          minimum: 0
          type: integer
          description: The number of spots reserved for specific uses
          format: int32
          nullable: true
      additionalProperties: false
      description: Data for updating specific fields of an existing parking lot.
...
    ReadHotelPassDto:
      required:
      - end
      - extraTime
      - id
      - licensePlate
      - parkingLotId
      - start
      type: object
      properties:
        id:
          type: integer
          description: Unique pass ID
          format: int64
        licensePlate:
          minLength: 1
          type: string
          description: Guest license plate
        parkingLotId:
          type: integer
          format: int64
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        extraTime:
          type: string
          format: date-span
      additionalProperties: false
      description: Details of a Hotel Pass.
...
    ReadParkingLotDto:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the parking lot
          format: int64
        name:
          type: string
          description: The display name
          nullable: true
        location:
          type: string
          description: The general location
          nullable: true
        address:
          type: string
          description: The full address
          nullable: true
        reserved:
          type: integer
          description: Number of reserved spots
          format: int32
        capacity:
          type: integer
          description: Total capacity
          format: int32
        tariff:
          type: number
          description: Hourly tariff
          format: double
        dayTariff:
          type: number
          description: Daily tariff (if applicable)
          format: double
          nullable: true
        createdAt:
          type: string
          description: Date of creation
          format: date-time
        status:
          $ref: '#/components/schemas/ParkingLotStatus'
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/ReservationModel'
          description: Associated reservations
          nullable: true
      additionalProperties: false
      description: Details of a registered parking lot.
...
    ReservationModel:
      required:
      - cost
      - createdAt
      - endTime
      - licensePlateNumber
      - parkingLotId
      - startTime
      - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        licensePlateNumber:
          minLength: 1
          type: string
        licensePlate:
          $ref: '#/components/schemas/LicensePlateModel'
        parkingLotId:
          type: integer
          format: int64
        parkingLot:
          $ref: '#/components/schemas/ParkingLotModel'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ReservationStatus'
        createdAt:
          type: string
          format: date-time
        cost:
          type: number
          format: double
      additionalProperties: false
...
    ReservationStatus:
      enum:
      - 0
      - 1
      - 2
      - 3
      - 4
      type: integer
      format: int32
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Please enter a valid token (e.g., ''Bearer eyJ...'').

        You can obtain a token by logging in or registering a new account via the
        endpoints.'
      scheme: Bearer
      bearerFormat: JWT
security:
- Bearer: []